moved circ_dist to generic utils script. using scipy circmean in scoring function.
